Fix unresolved reference linking errors for libyaml
On systems that default to linking with "--as-needed" flag, references to
libyaml are unresolved. This is because the order of libraries on the
command line is important and they must come after any files that need
the library.

Drop the path '/usr/local/lib' as it is not correct at least for ubuntu.
Perhaps it is needed on MacOS, but if so it should be conditional.

Signed-off-by: Rob Herring <robh@kernel.org>